First crack at a class and a function for linked lists.
