+++ Listerine.cpp
@@ -1,18 +1,10 @@
 #include "Workbench.h"
-
-/*
- * Definition for singly-linked list. */
- struct ListNode {
-     int val;
-     ListNode *next;
-     ListNode() : val(0), next(nullptr) {}
-     ListNode(int x) : val(x), next(nullptr) {}
-     ListNode(int x, ListNode *next) : val(x), next(next) {}
- };
+#include "Listerine.h"
  
-class Solution {
-public:
-	ListNode* addTwoNumbers(ListNode* l1, ListNode* l2) {
+ListNode* Solution::addTwoNumbers(ListNode* l1, ListNode* l2) {
+	return l1;
+}
 
-	}
-};
+int Solution::showList(ListNode* l1) {
+	return l1->val;
+}
